CTPI-240 : Remove ProductTypeITUtils.java
